Started to implement unit tests for the static ValueConverter class.
